Changes
- Formatting of all classes
- Fixed lots of warnings (seriously guys)

+++ src/cu/edu/cujae/graphy/algorithms/DialShortestPath.java
@@ -22,11 +22,11 @@ import cu.edu.cujae.graphy.core.Edge;
 import cu.edu.cujae.graphy.core.WeightedGraph;
 import cu.edu.cujae.graphy.core.iterators.GraphIterator;
 import cu.edu.cujae.graphy.utils.Pair;
+import java.util.Deque;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
 import java.util.Map;
-import java.util.Deque;
 import java.util.TreeMap;
 
 /**
@@ -38,7 +38,8 @@ import java.util.TreeMap;
  *
  * @author Ananda
  */
-public class DialShortestPath extends AbstractAlgorithm<Map<Integer, Pair<Integer, List<Integer>>>> {
+public class DialShortestPath extends AbstractAlgorithm<Map<Integer, Pair<Integer, List<Integer>>>>
+{
 
     private final static int INFINITY = Integer.MAX_VALUE;
     private final WeightedGraph<?> graph;
@@ -47,17 +48,19 @@ public class DialShortestPath extends AbstractAlgorithm<Map<Integer, Pair<Intege
     private final Map<Integer, Integer> distances;
     private final Map<Integer, Integer> previous;
     private final int numberOfVertices;
-    private final int W;   
+    private final int W;
 
-    public DialShortestPath(WeightedGraph<?> graph, GraphIterator<?> s, int maxW) {
+    public DialShortestPath(WeightedGraph<?> graph, GraphIterator<?> s, int maxW)
+    {
         super(new HashMap<>(graph.size()));
         //debo verificar que ninguna arista posea un peso mayor que maxWeight!!!
-        if(!graph.isWeighted()){
+        if (!graph.isWeighted())
+        {
             throw new IllegalArgumentException(
                     "Attempted to apply Dial algorithm to an unweighted graph.");
         }
-        
-        this.buckets = new HashMap<>((int) maxW * graph.size() + 1);
+
+        this.buckets = new HashMap<>(maxW * graph.size() + 1);
         this.distances = new TreeMap<>();
         this.graph = graph;
         this.previous = new TreeMap<>();
@@ -66,50 +69,57 @@ public class DialShortestPath extends AbstractAlgorithm<Map<Integer, Pair<Intege
         this.source = s;
 
         //Inicializar las distancias a infinito.
-        for(int v : graph.getLabels()){
+        for (int v : graph.getLabels())
+        {
             distances.put(v, INFINITY);
             previous.put(v, null);
         }
 
         //Inicializar las cubetas.
-        for(int i = 0; i < (maxW * numberOfVertices + 1); ++i){
+        for (int i = 0; i < (maxW * numberOfVertices + 1); ++i)
+        {
             buckets.put(i, new LinkedList<>());
         }
-        
+
         // Set the distance to the source to 0
         // and put it in the 0 bucket
         distances.put(s.getLabel(), 0);
-        buckets.get(0).push(s.getLabel());        
+        buckets.get(0).push(s.getLabel());
 
         //System.out.println(distances);
         //System.out.println(buckets);
     }
 
     @Override
-    public Algorithm<Map<Integer, Pair<Integer, List<Integer>>>> apply() {
+    public Algorithm<Map<Integer, Pair<Integer, List<Integer>>>> apply()
+    {
 
         //Paso 1: Verificar el ìndice de las cubetas y si estàn vacías.
         int index = 0;
         while (true)
         {
-            while (buckets.get(index).isEmpty() && index < W * numberOfVertices){
+            while (buckets.get(index).isEmpty() && index < W * numberOfVertices)
+            {
                 index++;
             }
-            if (index == W * numberOfVertices){
+            if (index == W * numberOfVertices)
+            {
                 break;
             }
             GraphIterator<?> u = graph.iterator(buckets.get(index).pop());
-            
+
             // Paso 2: Procesar todos los adyacentes del vértice u y actualizar sus distancias si es requerido.
-            for (Edge e : u.getAllAdjacentEdges()){
+            for (Edge e : u.getAllAdjacentEdges())
+            {
                 int v = e.getFinalNode().getLabel();
                 int weight = (int) e.getWeight().getValue();
-                
+
                 int du = distances.get(u.getLabel());
                 int dv = distances.get(v);
-                
+
                 //Verificar si hay un camino corto de v a u. 
-                if (dv > du + weight){
+                if (dv > du + weight)
+                {
                     distances.put(v, du + weight);
                     dv = distances.get(v);
                     previous.put(v, u.getLabel());
@@ -118,7 +128,8 @@ public class DialShortestPath extends AbstractAlgorithm<Map<Integer, Pair<Intege
                 }
             }
         }
-        for (int v : distances.keySet()){
+        for (int v : distances.keySet())
+        {
             getResult().put(v, new Pair<>(distances.get(v), makeShortestPathSequence(source.getLabel(), v)));
         }
         return this;
